Chop off the '_id' when looking up the association.
This allows giving :foo_id as the attribute name, which lets the view
helpers work ("form.select :foo_id, ..." checks for errors on :foo_id,
not :foo).  It also fixes use of :allow_nil in rails 2.1, which used
to always let the validation pass if the option was given.

The '_id' is only chopped if it's present, so it should be
backwards-compatible.